refactored retry
