+++ rust/src/domain/register/steps/execute.rs
@@ -12,7 +12,8 @@ use crate::domain::common::{
 };
 use crate::quote_cache;
 
-use super::super::retry::with_retry;
+use crate::retry::with_retry;
+
 use super::super::wizard::{StepContext, StepResult, WizardState};
 
 pub(crate) async fn run(state: &mut WizardState, ctx: &StepContext) -> Result<StepResult> {
