+++ rust/src/domain/register/steps/review.rs
@@ -15,7 +15,8 @@ use crate::domain::common::{
 use crate::environments;
 use crate::quote_cache;
 
-use super::super::retry::with_retry;
+use crate::retry::with_retry;
+
 use super::super::wizard::{ContactsChoice, StepContext, StepResult, WizardState};
 
 pub(crate) async fn run(state: &mut WizardState, ctx: &StepContext) -> Result<StepResult> {
